Build fixes for AMD-64

+++ dyninstAPI/src/process.C
@@ -4698,7 +4698,7 @@ bool process::handleStopThread(EventRecord &ev)
     }
 
     mal_printf("handling stopThread %lx=>%lx %s[%d]\n", 
-               pointAddress, (int)calculation, FILE__,__LINE__); 
+               pointAddress, (long)calculation, FILE__,__LINE__); 
 /* 2. If the callbackID is negative, the calculation is meant to be
       interpreted as the address of code, so we call stopThreadCtrlTransfer
       to translate the target to an unrelocated address */
